Add logic to RunCodeCleanupCSharp
